chore(chat): sync prompts and conversations methods (Issue #792)
