+++ apps/chat/src/store/conversations/conversations.reducers.ts
@@ -9,7 +9,7 @@ import {
   getNextDefaultName,
 } from '@/src/utils/app/folders';
 import { getConversationRootId } from '@/src/utils/app/id';
-import { isEntityExternal } from '@/src/utils/app/share';
+import { isEntityOrParentsExternal } from '@/src/utils/app/share';
 import { translate } from '@/src/utils/app/translation';
 
 import {
@@ -18,7 +18,7 @@ import {
   LikeState,
   Message,
 } from '@/src/types/chat';
-import { UploadStatus } from '@/src/types/common';
+import { FeatureType, UploadStatus } from '@/src/types/common';
 import { FolderInterface, FolderType } from '@/src/types/folder';
 import { SearchFilters } from '@/src/types/search';
 import { PublishRequest } from '@/src/types/share';
@@ -33,10 +33,6 @@ import uniq from 'lodash-es/uniq';
 export { ConversationsSelectors };
 
 const initialState: ConversationsState = {
-  conversationsToMigrateCount: 0,
-  migratedConversationsCount: 0,
-  isChatsBackedUp: false,
-  failedMigratedConversations: [],
   conversations: [],
   selectedConversationsIds: [],
   folders: [],
@@ -63,51 +59,6 @@ export const conversationsSlice = createSlice({
   initialState,
   reducers: {
     init: (state) => state,
-    migrateConversationsIfRequired: (state) => state,
-    initConversationsMigration: (
-      state,
-      {
-        payload,
-      }: PayloadAction<{
-        conversationsToMigrateCount: number;
-      }>,
-    ) => {
-      state.conversationsToMigrateCount = payload.conversationsToMigrateCount;
-    },
-    migrateConversationFinish: (
-      state,
-      {
-        payload,
-      }: PayloadAction<{
-        migratedConversationsCount: number;
-      }>,
-    ) => {
-      state.migratedConversationsCount = payload.migratedConversationsCount;
-    },
-    setFailedMigratedConversations: (
-      state,
-      {
-        payload,
-      }: PayloadAction<{
-        failedMigratedConversations: Conversation[];
-      }>,
-    ) => {
-      state.failedMigratedConversations = payload.failedMigratedConversations;
-    },
-    setIsChatsBackedUp: (
-      state,
-      {
-        payload,
-      }: PayloadAction<{
-        isChatsBackedUp: boolean;
-      }>,
-    ) => {
-      state.isChatsBackedUp = payload.isChatsBackedUp;
-    },
-    skipFailedMigratedConversations: (
-      state,
-      _action: PayloadAction<{ idsToMarkAsMigrated: string[] }>,
-    ) => state,
     initSelectedConversations: (state) => state,
     getSelectedConversations: (state) => state,
     initFoldersAndConversations: (state) => state,
@@ -156,9 +107,10 @@ export const conversationsSlice = createSlice({
 
         return conv;
       });
+
       if (payload.newId !== payload.oldConversation.id) {
         state.selectedConversationsIds = state.selectedConversationsIds.map(
-          (cid) => (cid === payload.newId ? payload.oldConversation.id! : cid),
+          (id) => (id === payload.newId ? payload.oldConversation.id! : id),
         );
       }
     },
@@ -183,13 +135,14 @@ export const conversationsSlice = createSlice({
 
         return conv;
       });
+
       if (
         payload.id &&
         payload.conversation.id &&
         payload.id !== payload.conversation.id
       ) {
         state.selectedConversationsIds = state.selectedConversationsIds.map(
-          (cid) => (cid === payload.id ? payload.conversation.id! : cid),
+          (id) => (id === payload.id ? payload.conversation.id! : id),
         );
       }
     },
@@ -207,9 +160,7 @@ export const conversationsSlice = createSlice({
       state,
       { payload }: PayloadAction<{ conversationIds: string[] }>,
     ) => {
-      const newSelectedIds = uniq(payload.conversationIds);
-
-      state.selectedConversationsIds = newSelectedIds;
+      state.selectedConversationsIds = uniq(payload.conversationIds);
     },
     unselectConversations: (
       state,
@@ -304,6 +255,7 @@ export const conversationsSlice = createSlice({
       state.selectedConversationsIds = state.selectedConversationsIds.filter(
         (id) => !payload.deleteIds.has(id),
       );
+      state.conversationsLoaded = true;
     },
     uploadConversationsByIds: (
       state,
@@ -350,24 +302,18 @@ export const conversationsSlice = createSlice({
       state,
       _action: PayloadAction<{
         newConversation: Conversation;
-        idToReplaceWithNewOne?: string;
       }>,
     ) => state,
     saveNewConversationSuccess: (
       state,
       {
-        payload: { newConversation, idToReplaceWithNewOne },
+        payload: { newConversation },
       }: PayloadAction<{
         newConversation: Conversation;
-        idToReplaceWithNewOne?: string;
       }>,
     ) => {
       state.conversations = state.conversations.concat(newConversation);
-      state.selectedConversationsIds = idToReplaceWithNewOne
-        ? state.selectedConversationsIds.map((id) =>
-            id === idToReplaceWithNewOne ? newConversation.id : id,
-          )
-        : [newConversation.id];
+      state.selectedConversationsIds = [newConversation.id];
 
       state.areSelectedConversationsLoaded = true;
     },
@@ -402,12 +348,9 @@ export const conversationsSlice = createSlice({
         payload,
       }: PayloadAction<{
         conversations: ConversationInfo[];
-        ignoreCombining?: boolean;
       }>,
     ) => {
-      state.conversations = payload.ignoreCombining
-        ? payload.conversations
-        : combineEntities(state.conversations, payload.conversations);
+      state.conversations = payload.conversations;
       state.conversationsLoaded = true;
     },
     addConversations: (
@@ -435,21 +378,11 @@ export const conversationsSlice = createSlice({
       state.areSelectedConversationsLoaded = false;
     },
     clearConversationsSuccess: (state) => {
-      state.conversations = state.conversations.filter(
-        (conv) =>
-          isEntityExternal(conv) ||
-          ConversationsSelectors.hasExternalParent(
-            { conversations: state },
-            conv.folderId,
-          ),
+      state.conversations = state.conversations.filter((conv) =>
+        isEntityOrParentsExternal(state, conv, FeatureType.Chat),
       );
-      state.folders = state.folders.filter(
-        (folder) =>
-          isEntityExternal(folder) ||
-          ConversationsSelectors.hasExternalParent(
-            { conversations: state },
-            folder.folderId,
-          ),
+      state.folders = state.folders.filter((folder) =>
+        isEntityOrParentsExternal(state, folder, FeatureType.Chat),
       );
     },
     createFolder: (
@@ -517,7 +450,6 @@ export const conversationsSlice = createSlice({
     deleteAllTemporaryFolders: (state) => {
       state.temporaryFolders = [];
     },
-
     renameTemporaryFolder: (
       state,
       { payload }: PayloadAction<{ folderId: string; name: string }>,
@@ -573,7 +505,7 @@ export const conversationsSlice = createSlice({
       state,
       { payload }: PayloadAction<{ folders: FolderInterface[] }>,
     ) => {
-      state.folders = combineEntities(payload.folders, state.folders);
+      state.folders = combineEntities(state.folders, payload.folders);
     },
     setSearchTerm: (
       state,
@@ -714,14 +646,13 @@ export const conversationsSlice = createSlice({
     playbackCancel: (state) => {
       state.isPlaybackPaused = true;
     },
-
     uploadConversationsWithFolders: (
       state,
-      { payload }: PayloadAction<{ paths: (string | undefined)[] }>,
+      { payload }: PayloadAction<{ ids: string[] }>,
     ) => {
       state.foldersStatus = UploadStatus.LOADING;
       state.loadingFolderIds = state.loadingFolderIds.concat(
-        payload.paths as string[],
+        payload.ids as string[],
       );
     },
     uploadFoldersSuccess: (
@@ -850,6 +781,33 @@ export const conversationsSlice = createSlice({
       state.chartLoading = false;
     },
     cleanupIsolatedConversation: (state) => state,
+    uploadChildConversationsWithFoldersSuccess: (
+      state,
+      {
+        payload,
+      }: PayloadAction<{
+        parentIds: string[];
+        folders: FolderInterface[];
+        conversations: ConversationInfo[];
+      }>,
+    ) => {
+      state.loadingFolderIds = state.loadingFolderIds.filter(
+        (id) => !payload.parentIds.includes(id),
+      );
+      state.folders = combineEntities(
+        state.folders,
+        payload.folders.map((folder) => ({
+          ...folder,
+          status: payload.parentIds.includes(folder.id)
+            ? UploadStatus.LOADED
+            : undefined,
+        })),
+      );
+      state.conversations = combineEntities(
+        state.conversations,
+        payload.conversations,
+      );
+    },
   },
 });
