+++ apps/chat/src/store/prompts/prompts.selectors.ts
@@ -319,14 +319,6 @@ export const isPromptLoading = createSelector([rootSelector], (state) => {
   return state.isPromptLoading;
 });
 
-export const selectPromptsToMigrateAndMigratedCount = createSelector(
-  [rootSelector],
-  (state) => ({
-    promptsToMigrateCount: state.promptsToMigrateCount,
-    migratedPromptsCount: state.migratedPromptsCount,
-  }),
-);
-
 // default name with counter
 export const selectNewFolderName = createSelector(
   [
@@ -341,19 +333,9 @@ export const selectNewFolderName = createSelector(
   },
 );
 
-export const selectFailedMigratedPrompts = createSelector(
-  [rootSelector],
-  (state) => state.failedMigratedPrompts,
-);
-
 export const selectIsNewPromptCreating = createSelector(
   [rootSelector],
   (state) => state.isNewPromptCreating,
-);
-
-export const selectIsPromptsBackedUp = createSelector(
-  [rootSelector],
-  (state) => state.isPromptsBackedUp,
 );
 
 export const getNewPrompt = createSelector([selectPrompts], (prompts) => {
