+++ apps/chat/src/store/prompts/prompts.types.ts
@@ -1,12 +1,8 @@
 import { FolderInterface } from '@/src/types/folder';
-import { Prompt, PromptInfo } from '@/src/types/prompt';
+import { PromptInfo } from '@/src/types/prompt';
 import { SearchFilters } from '@/src/types/search';
 
 export interface PromptsState {
-  migratedPromptsCount: number;
-  promptsToMigrateCount: number;
-  isPromptsBackedUp: boolean;
-  failedMigratedPrompts: Prompt[];
   prompts: PromptInfo[];
   folders: FolderInterface[];
   temporaryFolders: FolderInterface[];
