+++ apps/chat/src/store/settings/settings.epic.ts
@@ -24,6 +24,7 @@ import { errorsMessages } from '@/src/constants/errors';
 import { AddonsActions } from '../addons/addons.reducers';
 import { AuthSelectors } from '../auth/auth.reducers';
 import { ConversationsActions } from '../conversations/conversations.reducers';
+import { MigrationActions } from '../migration/migration.reducers';
 import { ModelsActions } from '../models/models.reducers';
 import { PromptsActions } from '../prompts/prompts.reducers';
 import { ShareActions } from '../share/share.reducers';
@@ -55,8 +56,7 @@ const initEpic: AppEpic = (action$, state$) =>
               );
               const actions: Observable<AnyAction>[] = [
                 of(UIActions.init()),
-                of(ConversationsActions.migrateConversationsIfRequired()),
-                of(PromptsActions.migratePromptsIfRequired()),
+                of(MigrationActions.init()),
                 of(ModelsActions.init()),
                 of(AddonsActions.init()),
                 of(ConversationsActions.init()),
